% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_na}
\alias{check_na}
\title{Check misssing values in variable by geo and time}
\usage{
check_na(data, var)
}
\arguments{
\item{data}{A data.frame}

\item{var}{A name of column to check}
}
\description{
Check misssing values in variable by geo and time
}
\examples{
  check_na(data_main_annual, nulc_aper)

}
