% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{EUR}
\alias{EUR}
\title{calculate EUR values from other currencies}
\usage{
EUR(x, time, currency, eur_rate)
}
\arguments{
\item{x}{a numeric vector to turn to euros.}

\item{time}{a time (year) vector.}

\item{currency}{a vector of currencies.}

\item{eur_rate}{a data.frame of exchange rates}
}
\description{
calculate EUR values from other currencies
}
