% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imput.R
\name{imput_change}
\alias{imput_change}
\title{Imput level series with change}
\usage{
imput_change(x, d)
}
\arguments{
\item{x}{A numeric vector to imput. NA values are imputed.}

\item{d}{A numeric vector of percentage change}
}
\description{
Imput level series with change
}
\examples{
t <- data.frame(a = c(NA, 2, NA, NA), b = c(0.7, 0.3, 0.1, 0.2))
dplyr::mutate(t, c = imput_change(a, b))

}
