% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ind_ulc}
\alias{ind_ulc}
\title{Nominal unit labour costs}
\usage{
ind_ulc(cost, output, input1 = 1, input2 = 1, time, baseyear)
}
\arguments{
\item{cost}{Compensation of employees}

\item{output}{value added or gdp}

\item{input1}{a labour input of costs (not neccessary if input = input2)}

\item{input2}{a labour input of output (not neccessary if input = input2)}
}
\description{
Nominal unit labour costs
}
\examples{
ind_ulc(cost = c(1,2,3), output = c(NA,4,6), time = c(1,2,3), baseyear = 2)
}
