% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ameco.R
\name{read_ameco}
\alias{read_ameco}
\title{Read Ameco files}
\usage{
read_ameco(table_num, file = NULL)
}
\arguments{
\item{table_num}{A link to a zip file.}
}
\description{
Gives warning because first line is longer than rest (one empty).
}
