% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_board_figs}
\alias{save_board_figs}
\title{Save figures for competitiveness board report}
\usage{
save_board_figs(
  filename,
  plot = last_plot(),
  width = 13.5,
  height = 13.5,
  save_data = FALSE
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width}{Plot size}

\item{height}{Plot size}

\item{save_data}{A logical to save plot data.}
}
\description{
to Downloads folder
}
\details{
Uses ggsave.
}
