% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_fig}
\alias{save_fig}
\title{Save figures for report}
\usage{
save_fig(filename, plot = last_plot(), width = 5.8, height = 3.9)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width}{Plot size}

\item{height}{Plot size}
}
\description{
Uses ggsave.
}
