% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{simple_lab}
\alias{simple_lab}
\title{Simplify labels (, or all vectors)}
\usage{
simple_lab(x, parts = 1, pattern = ", ")
}
\arguments{
\item{x}{A vector to simplify.}

\item{parts, }{A position of parts to include. Default 1.}

\item{pattern}{A  ", " Pattern to look for.}
}
\description{
Split based on a pattern and includes selected parts
}
\examples{
simple_lab(c(nulc = "unit labour cost, nominal"), parts = 1)
simple_lab(x = c(nulc = "unit labour cost, nominal", nulc_va = "unit labour cost, nominal, value added"), parts = c(2,3))
simple_lab(x = as.factor(c(nulc = "unit labour cost, nominal", nulc_va = "unit labour cost, nominal, value added")), parts = c(2,3))
simple_lab(var_labels_fi[c(36:41)], parts = c(3,4))

}
