% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate columns from code to names in data.frame}
\usage{
translate(.tbl, .var, trans_vec, simple = FALSE, ...)
}
\arguments{
\item{.tbl}{A tbl object.}

\item{.var}{A unquoted column name to translate.}

\item{trans_vec}{A named tranlation vector. Codes as names.}

\item{simple}{if TRUE simplyfied by \code{\link{simple_lab}}.
By default only first part of the name is taken, split by ",".}

\item{...}{further arguments to \code{\link{simple_lab}}.}
}
\description{
Translate columns from code to names in data.frame
}
\examples{
  tbl <- tibble::tibble(geo = c("FI", "SE"))
  t_geo <- c(FI = "Suomi", SE = "Ruotsi")
  translate(tbl, geo, t_geo)
  translate(tibble::tibble(vars = c("nulc", "nulc_va", "nulc_hw", "nulc_hw_va", "nulc_aper", "nulc_aper_va")),
            vars, var_labels_fi, simple = TRUE, parts = c(3, 4))

}
