% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighted_gmean}
\alias{weighted_gmean}
\title{Weighted geometric mean}
\usage{
weighted_gmean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{w}{a numeric vector for weights}

\item{na.rm}{A logical. Should missing x values be removed?}
}
\description{
Calculate weighted geometric mean
}
\examples{
x <- c(1,2,3, NA)
w <- c(0.25,0.5,0.25, NA)
weighted_gmean(x, w, na.rm = TRUE)
}
